+++ source/ci_source/ci_source_helpers.ts
@@ -9,7 +9,7 @@ import * as find from "lodash.find"
  * @param {[string]} keys Keys to ensure existence of
  * @returns {bool} true if they exist, false if not
  */
-export function ensureEnvKeysExist(env: Env, keys: Array<string>): boolean {
+export function ensureEnvKeysExist(env: Env, keys: string[]): boolean {
   /*const hasKeys = keys.map((key: string): boolean => {
     return env.hasOwnProperty(key) && env[key] != null && env[key].length > 0
   });
@@ -25,7 +25,7 @@ export function ensureEnvKeysExist(env: Env, keys: Array<string>): boolean {
  * @param {[string]} keys Keys to ensure existence and number-ness of
  * @returns {bool} true if they are all good, false if not
  */
-export function ensureEnvKeysAreInt(env: Env, keys: Array<string>): boolean {
+export function ensureEnvKeysAreInt(env: Env, keys: string[]): boolean {
   /*const hasKeys = keys.map((key: string): boolean => {
     return env.hasOwnProperty(key) && !isNaN(parseInt(env[key]))
   })